fix(report): keep the no-follow open compiling on every release platform
The release build cross-compiles for six platforms and Windows has no syscall.O_NOFOLLOW, so the previous commit broke that build. Local preflight builds only this host, and the six-platform build reached this branch when the release work merged, which is what surfaced it.

The open is now per-platform. POSIX keeps the no-follow and non-blocking flags. Windows has neither those flags nor FIFOs, so it checks the type and opens, and the comment states plainly that this leaves a window the standard library cannot close rather than implying parity. The FIFO test moved to a unix-tagged file so the test tree stays portable too.
